Unit Tests & Bugfixes
- Updated gitignore with json, csv, png
- Added unit testing for generate receipts
- Altered filename to standard python convention: Receipt-Analysis-Generate-Receipts.py --> receipt_analysis_generate_receipts.py
