+++ .gitignore
@@ -1,3 +1,7 @@
 # This .gitignore file should be placed at the root of your project directory
 #
 .env
+*.json
+*.csv
+*.png
+*.pyc
